requirements.txt
a2wsgi==1.10.10
aiologic==0.14.0
aiosqlite==0.21.0
alembic==1.16.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
apache-airflow==3.0.2
apache-airflow-client==3.0.2
apache-airflow-core==3.0.2
apache-airflow-providers-common-compat==1.7.1
apache-airflow-providers-common-io==1.6.0
apache-airflow-providers-common-sql==1.27.2
apache-airflow-providers-postgres==6.2.0
apache-airflow-providers-smtp==2.1.0
apache-airflow-providers-standard==1.3.0
apache-airflow-task-sdk==1.0.2
apispec==6.8.2
appnope==0.1.4
argcomplete==3.6.2
asgiref==3.8.1
asttokens==3.0.0
asyncpg==0.30.0
attr==0.3.2
attrs==25.3.0
babel==2.17.0
backcall==0.2.0
blinker==1.9.0
cachetools==6.1.0
cadwyn==5.4.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
colorlog==6.9.0
comm==0.2.2
configparser==7.2.0
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==44.0.1
Cython==3.1.2
decorator==5.2.1
Deprecated==1.2.18
dill==0.4.0
dl==0.1.0
dnspython==2.7.0
docutils==0.21.2
email_validator==2.2.0
executing==2.2.0
fastapi==0.115.13
fastapi-cli==0.0.7
filelock==3.18.0
Flask==2.3.3
Flask-AppBuilder==4.7.0
Flask-Babel==2.0.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.70.0
graphviz==0.21
greenlet==3.2.3
grpcio==1.73.0
gunicorn==23.0.0
h11==0.16.0
HTMLParser==0.0.2
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
ini2toml==0.15
ipython==8.12.3
ipywidgets==8.1.7
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyterlab_widgets==3.0.15
lazy-object-proxy==1.11.0
libcst==1.8.2
limits==5.4.0
linkify-it-py==2.0.3
lockfile==0.12.2
Mako==1.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
matplotlib-inline==0.1.7
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.7.0
msgspec==0.19.0
narwhals==1.43.1
numpy==2.3.0
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
ordered-set==4.1.0
packaging==25.0
pandas==2.3.0
pandas-stubs==2.2.3.250527
parso==0.8.4
pathspec==0.12.1
pendulum==3.1.0
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.2.1
pluggy==1.6.0
prison==0.2.1
prompt_toolkit==3.0.51
protobuf==5.29.5
psutil==7.0.0
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg-c==3.2.9
psycopg-pool==3.2.6
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
retryhttp==1.3.3
rich==13.9.4
rich-argparse==1.7.1
rich-toolkit==0.14.7
rpds-py==0.25.1
scikit-learn==1.7.0
scipy==1.15.3
setproctitle==1.3.6
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
streamlit==1.46.0
structlog==25.4.0
svcs==25.1.0
tabulate==0.9.0
tenacity==9.1.2
termcolor==3.1.0
text-unidecode==1.3
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
traitlets==5.14.3
typer==0.16.0
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
uc-micro-py==1.0.3
universal_pathlib==0.2.6
urllib3==2.5.0
uuid6==2025.0.0
uvicorn==0.34.3
uvloop==0.21.0
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
widgetsnbextension==4.0.14
wirerope==1.0.0
wrapt==1.17.2
WTForms==3.2.1
zipp==3.23.0